Move component styles to separate modules
