+++ src/components/contact/_story.stories.js
@@ -1,5 +1,5 @@
 import React from "react"
-// import { select } from "@storybook/addon-knobs"
+import { select } from "@storybook/addon-knobs"
 import Component from "./"
 
 // Category|Group/Sub-Group
